requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
mysqlclient==2.0.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
